[bigint] Move the magnitude to 64-bit limbs (#286)
`BigInt` kept its magnitude in base 2^32 but did all its arithmetic in
64-bit registers, so schoolbook multiplication and Knuth D each made
twice the passes they needed. This moves the words to base 2^64.

Speedup over the 32-bit build, same machine and benchmark:

| digits | 10 | 100 | 1,000 | 10,000 | 100,000 |
|---|---|---|---|---|---|
| add | 1.40x | 1.00x | 1.05x | 1.11x | 1.08x |
| multiply | 1.10x | 1.38x | 1.18x | 1.14x | 1.04x |
| floor divide | 1.99x | 1.32x | 1.86x | 1.22x | 1.10x |
| sqrt | 1.76x | 1.26x | 1.41x | 1.28x | 1.15x |

Add does not move because the kernels already read two 32-bit words as
one 64-bit limb. The million-digit rows do not move because the
transform packs bits, not words.

Against GMP, floor divide at 1000 digits goes from 2.87x slower to
1.52x, and at 10 digits from parity to 1.82x faster. sqrt at 1000
digits, 8.14x to 5.96x.

## The code

The three word kernels lose their pairing and their odd-length tails.
The packed base-2^64 schoolbook and the unpacked 32-bit one become one
Comba kernel, and `CUTOFF_PACK_64_AREA` goes with them -- there is
nothing left to pack.

Knuth D's quotient estimate now uses the Moller-Granlund reciprocal. The
last PR measured that as a loss and said not to retry it. That was a
32-bit finding: 64-by-32 is one `UDIV`, and there is no 128-by-64
instruction, so at this width the alternative is a call to a software
helper. `todo.md` says so now.

Decimal conversion uses 10^18 chunks. 10^19 fits a word, but 10^18 is
exactly twice what 10^9 carried in a half-width word, so the density per
byte is unchanged and the nine-digit grouping survives -- which matters
because `to_biguint()` hands the chunks straight to `BigUInt`, and only
whole nine-digit groups split cleanly.

`WordList` takes a word-type parameter; `BigUInt` stays on `uint32`.

Every cutoff was re-measured, since one counted in words now means twice
the value: `CUTOFF_KARATSUBA` 256 -> 64, `CUTOFF_TOOM3` 768 -> 256, the
NTT cost constant 1.10 -> 0.34 (crossover about 8000 words -> about
2300), `INLINE_WORDS` 12 -> 7, `CUTOFF_SQRT_BASE` 16 -> 32.
Burnikel-Ziegler and `CUTOFF_SQRT_RECURSIVE` did not move.

## What went wrong

All of this compiled cleanly and was correct at 32 bits.

`mag >>= 64`, peeling an `Int` into words. Shifting a 64-bit value by 64
is undefined, and arm64 takes the shift amount modulo the width, so it
does nothing and the loop appends a word forever. `BigInt(1) + 1` was
enough to exhaust memory and get the process killed. A word holds any
`Int`, so there is no loop to write.

`to_biguint()` reads the decimal chunks as `BigUInt` words. That is why
the chunk base cannot be 10^19.

`String(Int(words[0]))` formats through a signed type, so every value at
or above 2^63 printed negative. 2^64 - 1 came out as -1.

`borrow = (diff >> 63) & 1` took the borrow from the sign bit of a
widened 32-bit subtraction. At 64 bits a word of exactly 2^63 sets that
bit with no borrow at all.

Toom-3's exact division by three carried `(2^32 - 1) / 3`, and its
halving shifted the carry in at bit 31.

`product_range()` also lost a bound that was doing two jobs. `high <=
2^32 - 1` was about each factor fitting a word, which any non-negative
`Int` now does, but it was also keeping absurd ranges from running
forever. It caps the factor count instead.

## Checking

1076 Mojo tests and 58 CLI tests pass.

300 operand pairs checked against CPython rather than against an
internal identity: add, subtract, multiply, floor divide, modulo, both
shifts, and/or/xor, invert, decimal round-trip, sqrt. The widths
straddle every word, inline-capacity and cutoff boundary, and it runs
with `-D ASSERT=all` as well as without. Multiplication is also verified
by `(a+b)^2 - (a-b)^2 == 4ab` from 10 digits to 10^6, across schoolbook,
Karatsuba, Toom-3 and the transform.

## Breaking

`BigInt.words` is `UInt64` and the base is 2^64. A magnitude literal is
`[UInt64(1)]`. Anywhere your code took a word to be 32 bits it now has
to take it as 64: shifting by 32 becomes shifting by 64, masking with
`0xFFFF_FFFF` becomes masking with `0xFFFF_FFFF_FFFF_FFFF` or dropping
the mask, and a word count derived as `(bits + 31) // 32` becomes `(bits
+ 63) // 64`. Values and results are unchanged; only the representation
moved. `BigUInt` is untouched.

## Still open

sqrt at 1000 digits (5.96x) and the million-digit rows. Neither has a
limb width to blame any more. Ranked in `docs/internal/todo.md`.

+++ tests/bigint/test_bigint_basic.mojo
@@ -42,7 +42,7 @@ def test_from_int() raises:
         "BigInt(-9876543210): " + String(large_neg),
     )
 
-    # Test value that spans exactly 2 UInt32 words
+    # A value past the old 32-bit word boundary; one word now
     var two_words = BigInt(0xFFFF_FFFF + 1)  # 2^32 = 4294967296
     assert_true(
         String(two_words) == "4294967296",
@@ -174,13 +174,13 @@ def test_copy() raises:
 
 def test_normalize() raises:
     """Test _normalize strips leading zeros and normalizes -0."""
-    var x = BigInt(raw_words=[UInt32(42), UInt32(0), UInt32(0)], sign=False)
+    var x = BigInt(raw_words=[UInt64(42), UInt64(0), UInt64(0)], sign=False)
     x._normalize()
     assert_true(len(x.words) == 1, "normalize should strip leading zeros")
     assert_true(x.words[0] == 42, "normalize should preserve value")
 
     # Test -0 normalization
-    var neg_zero = BigInt(raw_words=[UInt32(0)], sign=True)
+    var neg_zero = BigInt(raw_words=[UInt64(0)], sign=True)
     neg_zero._normalize()
     assert_true(not neg_zero.sign, "-0 should normalize to +0")
